Added persistence flag and config doc update
The persistence forecast experiments were done by commenting out
the forecast step previously.  I switched this to a flag in the
configuration and changed the driver to conditionally forecast
based on this.  Moving forward, I should just create a specific
forecasting class that does nothing but persist the temps. Would
create more overhead in the runtime process though.

Added in some explanation of the configuration parameters that
was missing in this version.

Also added a small README describing basic usage of this code.