Add multivariate eofs to noise integration
Update the LMR_forecaster to use two stage EOF calculation in the
calibration of a LIM.  It will reduce the primary variable state
space, and then combine into a state, which is further reduced.

This provides consistency to the noise across the fields.  Prevents
unphysical perturbations between connected fields such as OHC
and surface temperature.