Flexible reconstruction output: scalar functions
Started the process of creating a more configuration-centric determination
of reconstruction outputs.  current plan is for scalar outputs for full
ensemble, and a mix of field outputs in an HDF5 file.  This commit contains
the creation of the scalar factor functions that will be mapped to
appropriate parts of the state.