Refactor of LMR_PSMbuild to use LMR_psms and config
PSMBuild still used old configuration style and reused the original
calibration in a series of complicated checks for doing an experiment
looking for objectively based PSMs.  This was a complicated feature
still under research and is not used in this online branch. Removed
the ability to do objective PSM determination and instead just use
the proxy metadata defined seasonality.  Now uses the new configuration
style to update and simplified the user inputs into this script. Also
now relies on the original Proxy and PSM classes to load and calibrate
everything.