Analysis loading functionality extensions
In order to handle on the fly calibration of PSMs using differing
seasonality and regridding targets, updates are made to the
GriddedVariable and Analysis variable to handle loading from
netCDF and pre-averaged files.  Updates include saving and averaging
for arbitrary season definitions, centering anomalies to specific
timer ranges, reducing data to specific time ranges, and usage of
the minimum fraction of non-NaN periods to constitute a valid average.