ForecastVariable generator and calib obj storage change
Usage of many different averaging periods in the prior state means there
are a lot of forecast data objs being created for the LIM, and these
were held on to at the detriment of memory.  Now stores necessary
information for compress/decompress of spatial fields and removes
the dobj after initialization.  Forecast variable loading is now
a generator for the load all method to facilitate on the fly
processing for the LIM state.