Update wrapper config
Update the wrapper configuration to have the working and archive directory
information, experiment name and handle single integer declarations for
the multi-seed.  Core configuration now defaults the working and experiment
name to the wrapper if not specified.  If a single integer is supplied
to multiseed it generates a random int list of seeds of length
of the iter range to use.  This prevents the user from having to provide
a long list.