Simplify scalar selection and make matrix factor functions
Simplify the function where scalar factors are selected to use a general
area average, single point, or pdo selection instead of a long and specific
list of different measures in an if statement.  Make functions available
to retrieve matrix factors instead of state object handling functions. This
is useful for multiplying reduced space output rather than having to
decompress the fields and then calculate a metric on that.