Handle edge trim when no ensemble specified
If no ensemble was specified the state loading function could not
concatenate the state due to different time periods covered. Alter
the load function slightly to take into account the years of the base
fields and pass them along to the psm required loading process so that
all fields when loaded have consistent times.