Add secondary NaN checks when working with diff model
When projecting a different model into the LIM space calibration
there is chance for leftover NaN values to blow the projection.
Added a second check during data compression to remove them when
going into LIM space.  The data will come out with the LIM field
forecast, I just have to adjust the calculation of the scalar factor
on the data.