Remove uncouple groups add long integration multiprocessing
Remove the uncouple groups option for the LIM and move to another
branch.  Add the option to run long integrations with a single process
with output for the arrays.  This was put in as an option to test whether
the integration on the ensemble year-by-year was affecting the noise stats.
It wasn't but I'm leaving the option in there.